% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy-fun.R
\name{dmy_hello}
\alias{dmy_hello}
\title{Dummy hello function}
\usage{
dmy_hello(who = "World", salutation = "Hello")
}
\arguments{
\item{who}{String target}

\item{salutation}{String kind}
}
\value{
A salutation string
}
\description{
Dummy hello function
}
\examples{
dmy_hello()
dmy_hello("Earth")
dmy_hello("Moon", "'Night")
}
