% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFNR-package.r
\docType{package}
\name{RFNR}
\alias{RFNR-package}
\alias{RFNR}
\title{RFNR: Recursive Filter-Augmented Nonlinear Regression (RFNR) Model}
\description{
Recursive Filter-Augmented Nonlinear Regression (RFNR) Model
Data Analysis for PERMANENT PEM fuel cell degradation:
A Stochastic Recursive Nonlinear Regression Model for
Recoverable Degradation in PEM Fuel Cells' by Parzer et. al.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RomanParzer/PERMANENT_DataAnalysis}
  \item Report bugs at \url{https://github.com/RomanParzer/PERMANENT_DataAnalysis/issues}
}

}
\author{
\strong{Maintainer}: Gianna Monti \email{gianna.monti@unimib.it}

Authors:
\itemize{
  \item Roman Parzer \email{TBD} [copyright holder]
}

}
\keyword{internal}
